+++ gw_full_latest/scripts/prepare_wifi_client.sh
@@ -1,16 +1,11 @@
 #!/bin/bash
 
-INTERFACES_NOT_AP_FILE=/etc/network/interfaces_not_ap
-INTERFACES_AP_FILE=/etc/network/interfaces_ap
-INTERFACES_FILE=/etc/network/interfaces
+pid = `pgrep hostapd`
 
-#if "interfaces_not_ap" exists, then replace current interfaces file by this one
-if [ -f $INTERFACES_NOT_AP_FILE ]; 
+if [ "$pid" == "" ]; 
   then
-	sudo mv $INTERFACES_FILE $INTERFACES_AP_FILE
-	sudo mv $INTERFACES_NOT_AP_FILE $INTERFACES_FILE
-	sudo systemctl disable hostapd.service
-	echo "The access point will be disabled at next reboot, using the file /etc/wpa_supplicant/wpa_supplicant.conf to connect to an access point."
-  else
 	echo "The access point is already disabled."
+  else	
+	sudo systemctl disable hostapd.service
+	echo "The access point will be disabled at next reboot, using the file /etc/wpa_supplicant/wpa_supplicant.conf to connect to an access point."	
 fi